코드 리팩토링 1차 완료
