+++ app/address/excel/entity.tsx
@@ -1,16 +1,16 @@
 import { User } from "next-auth";
 
+export interface CommonResponse<T> {
+    code: number;
+    message: string;
+    data: T;
+}
+
 export interface Groups {
     groupsId: number;
     userId: string;
     groupsName: string;
     parent: Groups | null;
-}
-
-export interface CommonResponse<T> {
-    code: number;
-    message: string;
-    data: T;
 }
 
 export interface Contact {
